djangoapp/requirements.txt
Django== 5.0.1,<5.1
django-cors-headers==4.3.1
psycopg2-binary== 2.9.9, <2.10
pillow == 10.2.0
django-redis==5.4.0
django-pandas==0.6.6
plotly==5.14.1
cx-Oracle==8.3.0
xlsxwriter==3.2.0
openpyxl==3.1.5
gunicorn==23.0.0
django-cleanup==8.1.0
django-tenants==3.7.0
django-baton==4.1.0
django-admin-charts==1.4.0